+++ sample-app/src/androidMain/kotlin/io/github/seyoungcho2/composeslider/ui/theme/Theme.android.kt
@@ -1,0 +1,40 @@
+package io.github.seyoungcho2.composeslider.ui.theme
+
+import android.app.Activity
+import android.os.Build
+import androidx.compose.material3.ColorScheme
+import androidx.compose.material3.dynamicDarkColorScheme
+import androidx.compose.material3.dynamicLightColorScheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.SideEffect
+import androidx.compose.ui.graphics.toArgb
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalView
+import androidx.core.view.WindowCompat
+
+
+@Composable
+actual fun getAppThemeColor(
+    darkTheme: Boolean,
+    // Dynamic color is available on Android 12+
+    dynamicColor: Boolean
+): ColorScheme {
+    val colorScheme = when {
+        dynamicColor && Build.VERSION.SDK_INT >= Build.VERSION_CODES.S -> {
+            val context = LocalContext.current
+            if (darkTheme) dynamicDarkColorScheme(context) else dynamicLightColorScheme(context)
+        }
+
+        darkTheme                                                      -> DarkColorScheme
+        else                                                           -> LightColorScheme
+    }
+    val view = LocalView.current
+    if (!view.isInEditMode) {
+        SideEffect {
+            val window = (view.context as Activity).window
+            window.statusBarColor = colorScheme.primary.toArgb()
+            WindowCompat.getInsetsController(window, view).isAppearanceLightStatusBars = darkTheme
+        }
+    }
+    return colorScheme
+}
