+++ sample-app/src/commonMain/kotlin/io/github/seyoungcho2/composeslider/ui/theme/Type.kt
@@ -1,22 +1,18 @@
 package io.github.seyoungcho2.composeslider.ui.theme
 
 import androidx.compose.material3.Typography
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.sp
-import io.github.seyoungcho2.composeslider.R
 
-val robotoFontFamily = FontFamily(
-    Font(R.font.roboto_regular, FontWeight.Normal, FontStyle.Normal)
-)
+
+expect fun getAppFontFamily(): FontFamily?
 
 // Set of Material typography styles to start with
 val Typography = Typography(
     bodyLarge = TextStyle(
-        fontFamily = robotoFontFamily,
+        fontFamily = getAppFontFamily(),
         fontWeight = FontWeight.Normal,
         fontSize = 16.sp,
         lineHeight = 24.sp,
