+++ sample-app/src/desktopMain/kotlin/io/github/seyoungcho2/composeslider/ui/theme/Type.desktop.kt
@@ -1,0 +1,6 @@
+package io.github.seyoungcho2.composeslider.ui.theme
+
+import androidx.compose.ui.text.font.FontFamily
+
+
+actual fun getAppFontFamily(): FontFamily? = null
