+++ sample-app/src/iosMain/kotlin/io/github/seyoungcho2/composeslider/ViewController.kt
@@ -1,0 +1,9 @@
+package io.github.seyoungcho2.composeslider
+
+import androidx.compose.ui.window.ComposeUIViewController
+import platform.UIKit.UIViewController
+
+
+fun MainViewController(): UIViewController = ComposeUIViewController {
+    App()
+}
